requirements.txt
streamlit
langchain
openai
python-docx
